+++ modules/swaggerAPI/init.js
@@ -34,7 +34,7 @@ exports.init = (app, apiUrl) => {
             "./modules/notification/notification-middleware/routes.js",
             "./modules/project-setting/routes.js",
             "./modules/trackerDownload/routes.js",
-            "./modules/taskIndex/routes.js",
+            "./modules/task-index/routes.js",
             "./modules/create-project/routes.js",
             "./modules/notification-count/routes.js",
             "./modules/trackerUserPermission/routes.js",
